raspi/requirements.txt
smbus2
retry
RPi.GPIO
python-dotenv
